+++ CHANGES.md
@@ -5,6 +5,7 @@ Version 3.1.0
 -------------
 - added CraftBukkitUpToDate
 - fixed color parsing (this may break 1.1 compatibility)
+- added a swear prevention
 
 Version 3.0.3
 -------------